vc/amd/agesa/f16kb: Cast to UINT64 to avoid overflow
Evaluated using 32-bit arithmetic, then used in a context that
expects an expression of type UINT64. Cast to UINT64 instead.

Found-by: Coverity CID 1241847
Reviewed-on: https://review.coreboot.org/c/coreboot/+/36081
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/vendorcode/amd/agesa/f16kb/Proc/Mem/Feat/DMI/mfDMI.c
@@ -716,7 +716,7 @@ MemFGetDctInterleavedLimit (
   UINT8 i;
 
   DctMemLimit = 0;
-  if (DctInterleavedMemSize == NBPtr->DCTPtr->Timings.DctMemSize << 6) {
+  if (DctInterleavedMemSize == (UINT64)(NBPtr->DCTPtr->Timings.DctMemSize) << 6) {
     // The whole memory range is interleaved for the DCTs with the minimum memory size
     for (i = 0; i < NBPtr->DctCount; i++) {
       DctMemLimit += DctInterleavedMemSize;
